Allow configuring username show at cursor
Introduces a --username command-line flag and the option to configure
the username via .teamtype/config, in order to have more flexibility on
a per project level without having to rely on Git.

Also comes with some small refactors of the code touched, e.g. adding a
consolidated ShareJoinFlags struct in the CLI code.